// src/commons/seguimirntoPlanes/EjeEstrategico/components/EjeEstrategico/BusquedaAvanzada/types.ts
export interface IBusquedaEjeEstrategico {
    id_eje_estrategico: number;
    nombre_plan:        string;
    nombre_tipo_eje:    string;
    nombre_programa:    string;
    nombre:             string;
    id_plan:            number;
    id_tipo_eje:        number;
    id_programa:        number;
}